+++ backend/src/main/java/com/example/cms/backup/ZipService.java
@@ -1,6 +1,6 @@
 package com.example.cms.backup;
 
-import com.example.cms.file.FileUtils;
+import com.example.cms.resource.FileUtils;
 import java.io.*;
 import java.nio.file.Path;
 import java.util.List;
