+++ backend/src/main/java/com/example/cms/resource/FileUtils.java
@@ -1,4 +1,4 @@
-package com.example.cms.file;
+package com.example.cms.resource;
 
 import java.io.File;
 import java.io.IOException;
