+++ frontend/modules/admin/src/lib/page/ui/page-resources/page-resources.component.html
@@ -54,7 +54,7 @@
 
         <td class="column-actions" *tuiCell="'actions'" tuiTd>
           <a
-            [routerLink]="['/files', item.id]"
+            [routerLink]="['/resources', item.id]"
             tuiIconButton
             size="s"
             appearance="flat"
