+++ frontend/modules/admin/src/lib/user/ui/user-resources/user-resources.component.html
@@ -53,7 +53,7 @@
 
         <td class="column-actions" *tuiCell="'actions'" tuiTd>
           <a
-            [routerLink]="['/files', item.id]"
+            [routerLink]="['/resources', item.id]"
             tuiIconButton
             size="s"
             appearance="flat"
